dash_template/requirements.txt
dash==2.6.0
pandas==1.4.2
sqlalchemy==1.4.36
langchain==0.1.0
openai==0.10.2
plotly==5.7.0
pytest==7.1.2
python-dotenv==0.19.1